#1039908 - Error in pedigree

+++ pedigree.php
@@ -124,7 +124,7 @@ for ($i=($controller->treesize-1); $i>=0; $i--) {
 		echo ($xoffset+$controller->pbwidth/2), 'px; top:', ($yoffset), 'px;">';
 		if ($can_go_back) {
 			$did = 1;
-			if ($i > ($controller->treesize/2) + ($controller->treesize/4)) {
+			if ($i > (int)($controller->treesize/2) + (int)($controller->treesize/4)) {
 				$did++;
 			}
 			echo '<a href=pedigree.php?PEDIGREE_GENERATIONS=', $controller->PEDIGREE_GENERATIONS, '&amp;rootid=', $controller->treeid[$did], '&amp;show_full=', $controller->show_full, '&amp;talloffset=', $controller->talloffset, ' class="icon-uarrow noprint"></a>';
@@ -149,7 +149,7 @@ for ($i=($controller->treesize-1); $i>=0; $i--) {
 	print_pedigree_person(WT_Person::getInstance($controller->treeid[$i]), 1, $iref, 1);		
 	if ($can_go_back) {
 		$did = 1;
-		if ($i > ($controller->treesize/2) + ($controller->treesize/4)) {
+		if ($i > (int)($controller->treesize/2) + (int)($controller->treesize/4)) {
 			$did++;
 		}
 		if ($TEXT_DIRECTION=="rtl") {$posn = 'right'; $arrow = 'icon-larrow';} else {$posn = 'left';	$arrow = 'icon-rarrow';	}
@@ -159,7 +159,7 @@ for ($i=($controller->treesize-1); $i>=0; $i--) {
 			echo '</div>';
 		} elseif ($talloffset < 2) {
 			echo '<div class="ancestorarrow" style="position:absolute; ',$posn,':', $controller->pbwidth+5, 'px; top:', ($controller->pbheight/2-10), 'px;">';
-				echo '<a href="pedigree.php?PEDIGREE_GENERATIONS='.$controller->PEDIGREE_GENERATIONS.'&amp;rootid='.$controller->treeid[$did].'&amp;show_full='.$controller->show_full.'&amp;talloffset='.$talloffset.'" class=" ',$arrow,' noprint"></a>';
+			echo '<a href="pedigree.php?PEDIGREE_GENERATIONS='.$controller->PEDIGREE_GENERATIONS.'&amp;rootid='.$controller->treeid[$did].'&amp;show_full='.$controller->show_full.'&amp;talloffset='.$talloffset.'" class=" ',$arrow,' noprint"></a>';
 			echo '</div>';
 		}
 	}
